+++ hospitapp/src/components/SearchFormClient.tsx
@@ -1,27 +1,46 @@
 "use client";
 
-import { FormEvent, useState } from "react";
-import { useRouter } from "next/navigation";
+import { FormEvent, useState, useEffect } from "react";
+import { useRouter, useSearchParams } from "next/navigation";
 import { SearchFormClientProps } from "@/services/search_ips/data_caching.service";
 import { SearchableSelect } from "./searchable_select";
+
+export type SearchFormSubmitHandler = (isSubmitting: boolean) => void;
 
 export default function SearchFormClient({
   specialties,
   eps,
-}: SearchFormClientProps) {
+  onSubmit,
+}: SearchFormClientProps & { onSubmit?: SearchFormSubmitHandler }) {
   const router = useRouter();
+  const searchParams = useSearchParams();
   const [_IS_SUBMITTING, set_is_submitting] = useState(false);
   const [_ERROR, set_error] = useState<string | null>(null);
+
+  const [initialSpecialties, setInitialSpecialties] = useState<string[]>(
+    searchParams.get("specialties")?.split(",").filter(Boolean) || []
+  );
+  const [initialEps, setInitialEps] = useState<string[]>(
+    searchParams.get("eps")?.split(",").filter(Boolean) || []
+  );
+
+  useEffect(() => {
+    const newSpecialties = searchParams.get("specialties")?.split(",").filter(Boolean) || [];
+    const newEps = searchParams.get("eps")?.split(",").filter(Boolean) || [];
+    setInitialSpecialties(newSpecialties);
+    setInitialEps(newEps);
+  }, [searchParams]);
 
   const _HANDLE_SUBMIT = async (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     set_is_submitting(true);
+    if (onSubmit) onSubmit(true);
     set_error(null);
 
     try {
       const _FORM_DATA = new FormData(e.currentTarget);
 
-      let coordinates: [number, number] = [-75.5849, 6.1816]; // Medellín Default Center
+      let coordinates: [number, number] = [-75.5849, 6.1816];
       try {
         const position = await new Promise<GeolocationPosition>(
           (resolve, reject) => {
@@ -40,10 +59,8 @@ export default function SearchFormClient({
         console.warn("No se pudo obtener la ubicación del usuario:", error);
       }
 
-      const max_distance = "20000"; // Default value of 20 km
-      const specialties = JSON.parse(
-        (_FORM_DATA.get("specialties") as string) || "[]"
-      );
+      const max_distance = "20000";
+      const specialties = JSON.parse((_FORM_DATA.get("specialties") as string) || "[]");
       const eps = JSON.parse((_FORM_DATA.get("eps") as string) || "[]");
       const page = "1";
       const page_size = "21";
@@ -57,41 +74,38 @@ export default function SearchFormClient({
         page_size,
       });
 
+      await new Promise((resolve) => setTimeout(resolve, 1000));
       router.push(`/results?${queryParams.toString()}`);
     } catch (err) {
       set_error(err instanceof Error ? err.message : "Unknown error occurred");
       console.error("Submission error:", err);
     } finally {
       set_is_submitting(false);
+      if (onSubmit) onSubmit(false);
     }
   };
 
   return (
     <div className="relative max-w-3xl w-full mx-auto px-6">
-      <form
-        onSubmit={_HANDLE_SUBMIT}
-        className=" p-8 rounded-2xl border-gray-100 space-y-6"
-      >
+      <form onSubmit={_HANDLE_SUBMIT} className="p-8 rounded-2xl border-gray-100 space-y-6">
         <div className="flex flex-col md:flex-row gap-6">
           <div className="flex-1">
-            <label className="block text-sm font-medium text-gray-700 mb-2">
-              EPS
-            </label>
+            <label className="block text-sm font-medium text-gray-700 mb-2">EPS</label>
             <SearchableSelect
               options={eps}
               placeholder="Selecciona EPS..."
               name="eps"
+              initialValues={initialEps}
             />
           </div>
 
           <div className="flex-1">
-            <label className="block text-sm font-medium text-gray-700 mb-2">
-              Especialidades
-            </label>
+            <label className="block text-sm font-medium text-gray-700 mb-2">Especialidades</label>
             <SearchableSelect
               options={specialties}
               placeholder="Selecciona especialidades..."
               name="specialties"
+              initialValues={initialSpecialties}
             />
           </div>
         </div>
@@ -105,7 +119,7 @@ export default function SearchFormClient({
         <button
           type="submit"
           disabled={_IS_SUBMITTING}
-          className={`w-full bg-blue-600 text-white text-lg font-semibold py-3 rounded-lg hover:bg-blue-700 transition-all duration-200 ease-in-out ${
+          className={`w-full bg-blue-600 text-white text-lg font-semibold py-3 rounded-lg hover:bg-blue-700 transition-all duration-100 ease-in-out ${
             _IS_SUBMITTING ? "cursor-wait opacity-75" : "cursor-pointer"
           }`}
         >
